MMM version 1.0.20

+++ scripts/powershell-modules/MdctecMaintenanceMenu/menu/02_ComplianceBase_Utilities/01_ComplianceBase_Installation/03_stop_and_remove_any_existing_containers.ps1
@@ -1,10 +1,21 @@
-function StopAndRemoveAnyExistingContainers{
-    param(
-        [Parameter(Mandatory=$true, HelpMessage='enter: $true or $false')]
-        $CreateDataBackup
-    )
+"Testing for existing container named 'cb-app'"
+if ((docker container inspect cb-app) -eq "[]"){
+    return
+}
+
+Import-Module MdctecMaintenanceMenu/support/SimplePrompt -DisableNameChecking -Force;
+
+Write-Warning "You're about to stop and remove the existing cb-app container."
+Write-Warning "You should be sure that no client data will be lost!"
+Write-Warning "If the data was mounted with a docker volume (which is the default behaviour) it should be safe to continue."
+Write-Host
+
+$AreYouSure = SimplePrompt "Type 'y' to continue, 'n' to abort" -Options y,n
+
+if ($AreYouSure -eq 'y'){
     "docker rm -f cb-app"
     docker rm -f cb-app;
 }
-
-StopAndRemoveAnyExistingContainers
+else{
+    "Aborted!"
+}
